requirements.txt
# 量化交易系统依赖
# 使用 pip install -r requirements.txt 安装

# ─── 数据处理 ─────────────────────────────────────────────────────────────────
pandas>=1.5.0,<3.0
numpy>=1.23.0,<2.0

# ─── 数据源 ───────────────────────────────────────────────────────────────────
tushare>=1.2.89
easyquotation>=0.7.5
baostock>=0.8.8        # 免费历史数据源（降级备选）
# pytdx>=1.72          # 通达信接口（可选，需要手动安装）
# mootdx>=0.9          # mootdx（可选）

# ─── Web框架 ──────────────────────────────────────────────────────────────────
flask>=2.3.0,<4.0

# ─── 数据可视化 ───────────────────────────────────────────────────────────────
plotly>=5.15.0

# ─── HTTP请求 ─────────────────────────────────────────────────────────────────
requests>=2.31.0

# ─── HTML解析 ─────────────────────────────────────────────────────────────────
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ─── 配置文件解析 ─────────────────────────────────────────────────────────────
pyyaml>=6.0

# ─── 工具库 ───────────────────────────────────────────────────────────────────
python-dateutil>=2.8.0

# ─── 定时任务 ─────────────────────────────────────────────────────────────────
apscheduler>=3.10.0
pytz>=2023.0

# ─── 测试 ─────────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-mock>=3.11.0
